# skip_whitespace: use a non-ref callback