@GuillaumeGomez 2025-06-20
Pretty neat, thanks!

I'm gonna send a follow-up to increase the version for a new release, unless you plan to send more before next release?

In any case, cc @Urgau