@GuillaumeGomez 2025-06-20
Then I'll wait for this new PR to be merged before making a new release. Can't wait to see what you'll come up with.