@hkBst 2025-06-20
@GuillaumeGomez If you're really curious, then see https://github.com/rust-lang/rust/pull/141001#issuecomment-2913226400.