@hkBst 2025-06-24
I think now that the perf wins here are a result of the simpler code triggering more inlining on ARM.